pyproject.toml
[build-system]
requires = ["hatchling>=1.22"]
build-backend = "hatchling.build"

[project]
name = "picture-annotator"
version = "0.1.0"
description = "Detect distant people in images (YOLO + SAHI) and export per-image bounding boxes."
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "sahi>=0.11.16,<0.12",
  "ultralytics>=8.1.0,<9",
  "tqdm>=4.66,<5",
  "pillow>=10.0,<12",
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]

[dependency-groups]
dev = [
  "pytest>=8,<9",
  "ruff>=0.6,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/picture_annotator"]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
